LP-78: Refactor notification
