+++ webapp/src/components/Context/UserContext.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const UserContext = React.createContext({})
+
+const UserProvider: React.FC = () => {
+  return <div></div>
+}
