+++ webapp/src/components/Layouts/NavbarLayout.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom'
 
-import { Navbar } from 'components/Navbar'
+import { Navbar } from 'components/common/Navbar'
 
 export const NavbarLayout: React.FC = () => {
   return (
